finish the “extract” part download frags in a random order split the code in different functions

+++ lib/Ocsinventory/Agent/Backend/Download.pm
@@ -12,7 +12,6 @@ use File::Path;
 use Archive::Extract;
 use File::Copy::Recursive qw(dirmove);
 
-
 use Data::Dumper;
 use Cwd;
 
@@ -24,6 +23,8 @@ sub clean {
     my $orderId = $params->{orderId};
     my $storage = $params->{storage};
 
+    my $cleanUpLevel = shift;
+
     my $downloadBaseDir = $config->{vardir}.'/download';
     my $targetDir = $downloadBaseDir.'/'.$orderId;
 
@@ -34,8 +35,12 @@ sub clean {
     my $level = [
 
     # Level 0
-    # only clean the run directory.
+    # only clean the part files
     sub {
+        my @part = glob("$targetDir/*.part");
+        return unless @part;
+
+        $logger->debug("Clean the partially downloaded files for $orderId");
         foreach (glob("$targetDir/*.part")) {
             if (!unlink($_)) {
                 $logger->error("Failed to clean $_ up");
@@ -46,7 +51,9 @@ sub clean {
     # Level 1
     # only clean the run directory.
     sub {
-        if (-d "$targetDir/run" && !rmtree("$targetDir/run")) {
+        return unless -d "$targetDir/run";
+        $logger->debug("Clean the $targetDir/run directory");
+        if (!rmtree("$targetDir/run")) {
             $logger->error("Failed to clean $targetDir/run up");
         }
     },
@@ -54,7 +61,10 @@ sub clean {
     # Level 2
     # clean the final file
     sub {
-        if (-f "$targetDir/final" && !unlink("$targetDir/final")) {
+        return unless -f "$targetDir/final";
+
+        $logger->debug("Clean the $targetDir/file file");
+        if (!unlink("$targetDir/final")) {
             $logger->error("Failed to clean $targetDir/final up");
         }
     },
@@ -62,23 +72,30 @@ sub clean {
     # Level 3
     # clean the PACK
     sub {
-        foreach (glob("$targetDir/*-*")) {
-            if (!unlink($_)) {
-                $logger->error("Failed to clean $_ up");
+        return unless -d $targetDir;
+
+        $logger->debug("Remove the fragment in $targetDir ");
+        if (!rmtree("$targetDir/run")) {
+                $logger->error("Failed to remove $targetDir");
             }
-        }
     },
 
 
     ];
 
+    if (!$cleanUpLevel || $cleanUpLevel >= @$level) {
+        $cleanUpLevel = @$level - 1;
+    }
+
+    foreach (0..$cleanUpLevel) {
+        $level->[$_]();
+    }
+
 }
 
-sub downloadAndExtract {
+sub extractArchive {
     my $params = shift;
 
-    use Data::Dumper;
-#    print Dumper($params);
     my $config = $params->{config};
     my $logger = $params->{logger};
     my $orderId = $params->{orderId};
@@ -88,113 +105,70 @@ sub downloadAndExtract {
 
     my $downloadBaseDir = $config->{vardir}.'/download';
     my $targetDir = $downloadBaseDir.'/'.$orderId;
-    if (!-d $targetDir && !mkdir ($targetDir)) {
-        $logger->error("Failed to create $targetDir");
+
+    return unless $order->{FRAGS}; # no frags → nothing to extract
+
+    if (!open FILE, "<$targetDir/final") {
+        $logger->error("Failed to open $targetDir/final: $!");
+        return;
     }
 
+    my $tmp;
+    read(FILE, $tmp, 16);
+    my $magicNumber = unpack("S<", $tmp);
 
-    if (!-f "$targetDir/run" && mkpath("$targetDir/run")) {
-        $logger->error("Failed to create $targetDir/run");
+    if (!$magicNumber) {
+        $logger->error("Failed to read magic number for $targetDir/final");
         return;
     }
 
 
-    if (!$order->{FRAGS}) {
-        $logger->info("No files to download/extract");
-    } else {
 
+    my $type = {
 
-        $logger->fault("order not correctly initialised") unless $order;
-        $logger->fault("config not correctly initialised") unless $config;
+        19280 => 'zip',
+        35615 => 'tgz', # well gzip...
 
-        $logger->debug("processing ".$orderId);
+    };
 
+    if (!$type->{$magicNumber}) {
+        $logger->error("Unknow magic number $magicNumber! ".
+            "Sorry I can't extract this archive ( $targetDir/final ). ".
+            "If you think, your archive is valide, please submit a bug on ".
+            "http://launchpad.net/ocsinventory with this message and the ".
+            "archive.");
+        return;
+    }
 
-        my $baseUrl = ($order->{PROTO} =~ /^HTTP$/i)?"http://":"";
-        $baseUrl .= $order->{PACK_LOC};
-        $baseUrl .= '/' if $order->{PACK_LOC} !~ /\/$/;
-        $baseUrl .= $orderId;
+    $Archive::Extract::DEBUG=$config->{debug}?1:0;
+    my $archiveExtract = Archive::Extract->new(
 
-        $logger->info("Download the file(s)");
-        # TODO randomise the order
-        foreach my $fragID (1..$order->{FRAGS}) {
-            my $frag = $orderId.'-'.$fragID;
+        archive => "$targetDir/final",
+        type => $type->{$magicNumber}
 
-            my $remoteFile = $baseUrl.'/'.$frag;
-            my $localFile = $targetDir.'/'.$frag;
+    );
 
-            next if -f $localFile; # Local file already here
+    if (!$archiveExtract->extract(to => "$targetDir/run")) {
+        $logger->error("Failed to extract archive $targetDir/run");
+        return;
+    }
 
-            my $rc = LWP::Simple::getstore($remoteFile, $localFile.'.part');
-            if (is_success($rc) && move($localFile.'.part', $localFile."/")) {
-                # TODO to a md5sum/sha256 check here
-                $logger->debug($remoteFile.' -> '.$localFile.': success');
+    $logger->debug("Archive $targetDir/run extracted");
 
-            } else {
-                $logger->error($remoteFile.' -> '.$localFile.': failed');
-                unlink ($localFile.'.part');
-                unlink ($localFile);
-                # TODO Count the number of failure
-                return;
-            }
-        }
+}
 
+sub processOrderCmd {
+    my $params = shift;
 
-        ### Recreate the archive
-        $logger->info("Construct the archive");
-        if (!open (FINALFILE, ">$targetDir/final")) {
-            $logger->error("Failed to open $targetDir/final");
-            return;
-        }
-        foreach my $fragID (1..$order->{FRAGS}) {
-            my $frag = $orderId.'-'.$fragID;
+    my $config = $params->{config};
+    my $logger = $params->{logger};
+    my $orderId = $params->{orderId};
+    my $storage = $params->{storage};
 
-            my $localFile = $targetDir.'/'.$frag;
-            if (!open (FRAG, "<$localFile")) {
-                $logger->error("Failed to open $localFile");
-                close FINALFILE;
-                $logger->error("Failed to remove $baseUrl") unless unlink $baseUrl;
-                return;
-            }
+    my $order = $storage->{byId}->{$orderId};
 
-            foreach (<FRAG>) {
-                if (!print FINALFILE) {
-                    # TODO, imagine a graceful clean up function
-                    $logger->error("Failed to open $localFile");
-                    close FINALFILE;
-                    $logger->error("Failed to remove $baseUrl") unless unlink $baseUrl;
-                    return;
-                }
-            }
-            close FRAG;
-        }
-        close FINALEFILE; # TODO catch the ret code
-
-        
-        # Turns debug mode on if needed
-        #$Archive::Extract::DEBUG=1 if $config->{debug};
-        # Prefere local binaries
-        $Archive::Extract::PREFER_BIN=1;
-
-        my $success = 0;
-        foreach my $type (qw/tgz zip tar tbz/) {
-            my $archive = Archive::Extract->new(
-                archive => "$targetDir/final",
-                type => $type);
-            if ($archive && $archive->extract(to => "$targetDir/run")) {
-                $logger->debug("Archive is type: $type");
-                $logger->info("Files extracted in $targetDir/run");
-                $success = 1; 
-                last;
-            }
-        }
-
-        if (!$success) {
-            $logger->error("Failed to extract $targetDir/final");
-            return;
-        }
-    } # No attach file to download/extract
-
+    my $downloadBaseDir = $config->{vardir}.'/download';
+    my $targetDir = $downloadBaseDir.'/'.$orderId;
 
     my $cwd = getcwd;
     if ($order->{ACT} eq 'EXECUTE') {
@@ -220,23 +194,147 @@ sub downloadAndExtract {
             }
         }
     } elsif ($order->{ACT} eq 'LAUNCH') {
-        my $cmd = $order->{'NAME'};
 
-        $logger->debug("Launching $cmd...");
+        my $cmd = $order->{'NAME'};
         if (!-f "$targetDir/run/$cmd") {
             $logger->error("$targetDir/run/$cmd not found");
             return;
         }
 
-        if (chmod(0755, "$targetDir/run/$cmd")) {
-            $logger->debug("Cannot chmod: $!");
+
+        if ($^O !~ /^MSWin/) {
+            $cmd .= './' unless $cmd =~ /^\//;
+            if (chmod(0755, "$targetDir/run/$cmd")) {
+                $logger->error("Cannot chmod: $!");
+            }
         }
+        
+        $logger->debug("Launching $cmd...");
+
         # TODO, add ./ only for non Windows OS.
-        chdir("$targetDir/run");
-        system( "./".$cmd );
-        chdir($cwd);
+        if (!chdir("$targetDir/run")) {
+            $logger->fault("Failed to chdir to $cwd");
+        }
+        system( $cmd );
+        if (!chdir($cwd)) {
+            $logger->fault("Failed to chdir to $cwd");
+        }
         # TODO, return the exit code
     }
+
+
+}
+
+sub downloadAndConstruct {
+    my $params = shift;
+
+    my $config = $params->{config};
+    my $logger = $params->{logger};
+    my $orderId = $params->{orderId};
+    my $storage = $params->{storage};
+
+    my $order = $storage->{byId}->{$orderId};
+
+    my $downloadBaseDir = $config->{vardir}.'/download';
+    my $targetDir = $downloadBaseDir.'/'.$orderId;
+
+    return unless $order->{FRAGS}; # no frags → nothing to extract
+
+
+    if (!$order->{FRAGS}) {
+        $logger->info("No files to download/extract");
+    } else {
+
+
+        $logger->fault("order not correctly initialised") unless $order;
+        $logger->fault("config not correctly initialised") unless $config;
+
+        $logger->debug("processing ".$orderId);
+
+
+        my $baseUrl = ($order->{PROTO} =~ /^HTTP$/i)?"http://":"";
+        $baseUrl .= $order->{PACK_LOC};
+        $baseUrl .= '/' if $order->{PACK_LOC} !~ /\/$/;
+        $baseUrl .= $orderId;
+
+        $logger->info("Download the file(s) if needed");
+
+
+        # Randomise the download order
+        my @downloadToDo;
+        foreach (0..($order->{FRAGS}-1)) {
+            push (@downloadToDo, 'a');
+        }
+        while (grep (/a/, @downloadToDo)) {
+
+            my $fragID = int(rand(@downloadToDo)); # pick a random frag
+            next unless $downloadToDo[$fragID]; # Already done?
+            $downloadToDo[$fragID] = undef;
+
+
+
+            my $frag = $orderId.'-'.$fragID;
+
+            my $remoteFile = $baseUrl.'/'.$frag;
+            my $localFile = $targetDir.'/'.$frag;
+
+            next if -f $localFile; # Local file already here
+
+            my $rc = LWP::Simple::getstore($remoteFile, $localFile.'.part');
+            if (is_success($rc) && move($localFile.'.part', $localFile)) {
+                # TODO to a md5sum/sha256 check here
+                $logger->debug($remoteFile.' -> '.$localFile.': success');
+
+            } else {
+                $logger->error($remoteFile.' -> '.$localFile.': failed');
+                unlink ($localFile.'.part');
+                unlink ($localFile);
+                # TODO Count the number of failure
+                return;
+            }
+        }
+
+
+        ### Recreate the archive
+        $logger->info("Construct the archive in $targetDir/final");
+        if (!open (FINALFILE, ">$targetDir/final")) {
+            $logger->error("Failed to open $targetDir/final");
+            return;
+        }
+        foreach my $fragID (1..$order->{FRAGS}) {
+            my $frag = $orderId.'-'.$fragID;
+
+            my $localFile = $targetDir.'/'.$frag;
+            if (!open (FRAG, "<$localFile")) {
+                $logger->error("Failed to open $localFile");
+                close FINALFILE;
+                $logger->error("Failed to remove $baseUrl") unless unlink $baseUrl;
+                return;
+            }
+
+            foreach (<FRAG>) {
+                if (!print FINALFILE) {
+                    # TODO, imagine a graceful clean up function
+                    $logger->error("Failed to write in $localFile: $!");
+                    clean(2);
+                    return;
+                }
+            }
+            close FRAG;
+        }
+        close FINALFILE; # TODO catch the ret code
+
+        
+        return unless extractArchive({
+            config => $config,
+            logger => $logger,
+            orderId => $orderId,
+            
+            });
+
+
+    } # No attach file to download/extract
+
 }
 
 
@@ -346,19 +444,50 @@ sub longRun {
     print "Storage".Dumper($storage);
 
     my $downloadBaseDir = $config->{vardir}.'/download';
-    if (!-f $downloadBaseDir && !mkpath($downloadBaseDir)) {
+    if (!-d $downloadBaseDir && !mkpath($downloadBaseDir)) {
         $logger->error("Failed to create $downloadBaseDir");
     }
 
     foreach my $priority (0..10) {
         foreach my $orderId (keys %{$storage->{byPriority}->[$priority]}) {
-            print $orderId."\n";
-            next unless downloadAndExtract({
+            clean({
+                    config => $config,
+                    logger => $logger,
+                    orderId => $orderId,
+                    storage => $storage,
+                },2);
+
+            my $targetDir = $downloadBaseDir.'/'.$orderId;
+            if (!-d "$targetDir/run" && !mkpath("$targetDir/run")) {
+                $logger->error("Failed to create $targetDir/run");
+                return;
+            }
+
+
+            downloadAndConstruct({
                     config => $config,
                     logger => $logger,
                     orderId => $orderId,
                     storage => $storage,
                 });
+            extractArchive({
+                    config => $config,
+                    logger => $logger,
+                    orderId => $orderId,
+                    storage => $storage,
+                });
+            processOrderCmd({
+                    config => $config,
+                    logger => $logger,
+                    orderId => $orderId,
+                    storage => $storage,
+                });
+            clean({
+                    config => $config,
+                    logger => $logger,
+                    orderId => $orderId,
+                    storage => $storage,
+                },2);
         }
     }
 }
